# bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PsychoHelp_API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/appointment/patients/{patientId}": {
      "get": {
        "tags": [
          "Appointment"
        ],
        "summary": "Retorna appointments por Id de patient",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "patientId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AppointmentResource"
              }
            }
          }
        }
      }
    },
    "/api/v1/appointment/psychologist/{psychoId}/patient/{patientId}": {
      "get": {
        "tags": [
          "Appointment"
        ],
        "summary": "Retorna todos los appointments de un psicólogo y un paciente",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "psychoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "patientId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AppointmentResource"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patients": {
      "get": {
        "tags": [
          "Patients"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PatientResource"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SavePatientResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Patients"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/patients/{id}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Patients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SavePatientResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/patients/email/{patientEmail}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "patientEmail",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AppointmentResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "patientId": {
          "format": "int32",
          "type": "integer"
        },
        "psychoId": {
          "format": "int32",
          "type": "integer"
        },
        "psychoNotes": {
          "type": "string"
        },
        "scheduleDate": {
          "format": "date-time",
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "motive": {
          "type": "string"
        },
        "personalHistory": {
          "type": "string"
        },
        "testRealized": {
          "type": "string"
        },
        "treatment": {
          "type": "string"
        }
      }
    },
...
    "PatientResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phone": {
          "format": "int64",
          "type": "integer"
        },
        "state": {
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "img": {
          "type": "string"
        }
      }
    },
...
    "SavePatientResource": {
      "required": [
        "date",
        "email",
        "firstName",
        "gender",
        "img",
        "lastName",
        "password",
        "phone",
        "state"
      ],
      "type": "object",
      "properties": {
        "firstName": {
          "maxLength": 50,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "type": "string"
        },
        "state": {
          "maxLength": 50,
          "type": "string"
        },
        "email": {
          "format": "email",
          "type": "string"
        },
        "password": {
          "maxLength": 20,
          "type": "string"
        },
        "phone": {
          "maxLength": 9,
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "img": {
          "type": "string"
        }
      }
    },
...
  }
}